Move BuildEndpoint to sender package
